Rename component imports
