+++ src/pages/contact.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import Button from "@components/Button";
-import Layout from "@components/Layout";
+import Button from "../components/Button";
+import Layout from "../components/Layout";
 
 const Contact = () => (
   <Layout title="Contact Us">
