+++ src/pages/index.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import styled from "styled-components";
 
 // Components
-import Layout from "@components/Layout";
-import Button from "@components/Button";
-import EventsDeck from "@components/EventsDeck";
-import Particles from "@components/Particles";
+import Layout from "../components/Layout";
+import Button from "../components/Button";
+import EventsDeck from "../components/EventsDeck";
+import Particles from "../components/Particles";
 
-import { color, breakpoint } from "@theme/config";
+import { color, breakpoint } from "../theme/config";
 
 const JoinButton = styled(Button)`
   font-size: 1.5rem;
