+++ src/pages/resources.tsx
@@ -3,9 +3,9 @@ import { graphql } from "gatsby";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import Layout from "@components/Layout";
+import Layout from "../components/Layout";
 
-import { color } from "@theme/config";
+import { color } from "../theme/config";
 
 const Title = styled.h1`
   margin-bottom: 0.5rem;
